+++ packages/components/src/components/date-picker/date-picker.tsx
@@ -42,6 +42,8 @@ if (
   customElements.define('duet-date-picker', DuetDatePickerCustomElement);
 }
 
+const DEFAULT_ICON_SIZE = 20;
+
 @Component({
   tag: 'scale-date-picker',
   shadow: false,
@@ -147,6 +149,8 @@ export class DatePicker {
   /** @deprecated */
   @Prop() size?: string;
 
+  @Prop() variant?: 'informational' | 'warning' | 'danger' | 'success' =
+    'informational';
   /** Whether the input element has focus */
   @State() hasFocus: boolean = false;
 
@@ -261,25 +265,29 @@ export class DatePicker {
     );
 
     if (calendarIcon) {
-      calendarIcon.replaceWith(
-        document.createElement('scale-icon-content-calendar')
-      );
+      const icon = document.createElement('scale-icon-content-calendar');
+      icon.size = DEFAULT_ICON_SIZE;
+      calendarIcon.replaceWith(icon);
     }
 
     const navLeftIcon = this.duetInput.querySelector('.duet-date__prev svg');
 
     if (navLeftIcon) {
-      navLeftIcon.replaceWith(
-        document.createElement('scale-icon-navigation-left')
+      const scaleNavLeftIcon = document.createElement(
+        'scale-icon-navigation-left'
       );
+      scaleNavLeftIcon.size = 16;
+      navLeftIcon.replaceWith(scaleNavLeftIcon);
     }
 
     const navRightIcon = this.duetInput.querySelector('.duet-date__next svg');
 
     if (navRightIcon) {
-      navRightIcon.replaceWith(
-        document.createElement('scale-icon-navigation-right')
+      const scaleNavRightIcon = document.createElement(
+        'scale-icon-navigation-right'
       );
+      scaleNavRightIcon.size = 16;
+      navRightIcon.replaceWith(scaleNavRightIcon);
     }
 
     const selectIcon = this.duetInput.querySelectorAll(
@@ -433,7 +441,6 @@ export class DatePicker {
   }
 
   render() {
-    const helperTextId = `helper-message-${this.internalId}`;
     return (
       <Host>
         {this.styles && <style>{this.styles}</style>}
@@ -444,7 +451,8 @@ export class DatePicker {
             this.invalid && `scale-date-picker--status-error`,
             this.hasFocus && 'scale-date-picker--focus',
             this.disabled && 'scale-date-picker--disabled',
-            this.hasValue && 'animated'
+            this.hasValue && 'animated',
+            this.helperText && 'has-helper-text'
           )}
         >
           <label class="date-picker__label" htmlFor={this.identifier}>
@@ -479,15 +487,11 @@ export class DatePicker {
               (this.duetInput = element)
             }
           ></duet-date-picker>
-          {!!this.helperText && (
-            <div
-              class="date-picker__meta"
-              id={helperTextId}
-              aria-live="polite"
-              aria-relevant="additions removals"
-            >
-              <div class="date-picker__helper-text">{this.helperText}</div>
-            </div>
+          {this.helperText && (
+            <scale-helper-text
+              helperText={this.helperText}
+              variant={this.invalid ? 'danger' : this.variant}
+            ></scale-helper-text>
           )}
         </div>
       </Host>