+++ packages/components/src/components/loading-spinner/loading-spinner.tsx
@@ -13,6 +13,7 @@ export class LoadingSpinner {
   @Prop() variant: 'white' | 'primary' = 'primary';
   @Prop() alignment: 'horizontal' | 'vertical' = 'horizontal';
   @Prop() text: string;
+  @Prop() accessibilityTitle: string;
   @Prop() size: 'small' | 'large' = 'small';
 
   componentWillLoad() {
@@ -29,7 +30,7 @@ export class LoadingSpinner {
                 class="path"
                 cx="25"
                 cy="25"
-                r="20"
+                r="22.5"
                 fill="none"
                 stroke-width="4"
               ></circle>
@@ -43,14 +44,16 @@ export class LoadingSpinner {
                 class="path"
                 cx="25"
                 cy="25"
-                r="20"
+                r="22.5"
                 fill="none"
                 stroke-width="4"
               ></circle>
             </svg>
           </div>
           <div class="sr-only" aria-live="polite" id={`spinner-label-${i}`}>
-            {this.text || 'Loading'}
+            {this.accessibilityTitle
+              ? this.accessibilityTitle
+              : this.text || 'Loading'}
           </div>
           {this.text ? (
             <div part="text" class="spinner__text" aria-hidden="true">