+++ packages/storybook-vue/stories/components/dropdown-select/ScaleDropdownSelect.vue
@@ -8,6 +8,8 @@
     :size="size"
     :value="value"
     :combobox-id="comboboxId"
+    :hide-label-visually="hideLabelVisually"
+    :floating-strategy="floatingStrategy"
     @scaleChange="['scale-change']"
     @scaleFocus="['scale-focus']"
     @scaleBlur="['scale-blur']"
@@ -30,6 +32,8 @@ export default {
     value: { type: String },
     variant: { type: String },
     comboboxId: { type: String },
+    hideLabelVisually: {type: Boolean, default: false},
+    floatingStrategy: { type: String }
   },
   methods: {
     'scale-change'($event) {